move base view-controller to architecture module and adjust some logic
