+++ app/src/main/java/com/kunminx/puremusic/ui/page/DrawerFragment.java
@@ -26,9 +26,8 @@ import androidx.annotation.Nullable;
 
 import com.kunminx.puremusic.BR;
 import com.kunminx.puremusic.R;
-import com.kunminx.puremusic.ui.base.BaseFragment;
-import com.kunminx.puremusic.ui.base.DataBindingConfig;
-import com.kunminx.puremusic.ui.callback.SharedViewModel;
+import com.kunminx.architecture.ui.page.BaseFragment;
+import com.kunminx.architecture.ui.page.DataBindingConfig;
 import com.kunminx.puremusic.ui.page.adapter.DrawerAdapter;
 import com.kunminx.puremusic.ui.state.DrawerViewModel;
